fix wallet settings and settings page conflict
